# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEO APP API",
    "version": "1.0.0",
    "description": "REST API ของร้าน NEO APP (ขายแอปพลิเคชัน + กระเป๋าเงิน)\n\n## การยืนยันตัวตน\n\nทุก endpoint ต้องผ่านการยืนยันตัวตน ไม่มีเส้นไหนเปิดสาธารณะ รองรับ 2 วิธี:\n\n1. **Cookie session (เบราว์เซอร์)** — ล็อกอินที่ `/login` แล้ว Supabase จะเก็บ JWT ไว้ใน cookie `sb-<project-ref>-auth-token`\n2. **Bearer token (Postman / curl / script)** — ส่งเฮดเดอร์ `Authorization: Bearer <access_token>`\n\nขอ access token:\n\n```bash\ncurl -X POST \"$NEXT_PUBLIC_SUPABASE_URL/auth/v1/token?grant_type=password\" \\\n  -H \"apikey: $NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY\" \\\n  -H \"content-type: application/json\" \\\n  -d '{\"email\":\"you@example.com\",\"password\":\"...\"}'\n```\n\n## สิทธิ์\n\nข้อมูลถูกคุมด้วย Row Level Security ของ Postgres: ลูกค้าเห็นเฉพาะคำสั่งซื้อ/กระเป๋าเงินของตัวเอง\nส่วนการเพิ่ม-แก้-ลบสินค้า อัปโหลดรูป และแก้ตั้งค่าร้าน ทำได้เฉพาะบัญชีที่มี `app_metadata.role = \"admin\"`\n\n## เอกสารนี้\n\nหน้า `/docs` และไฟล์ `/openapi.json` เปิดให้เฉพาะแอดมิน คนอื่นจะได้ 404",
    "contact": {
      "name": "NEO APP"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/orders/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "ORD-1024"
        }
      ],
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "ข้อมูลคำสั่งซื้อรายรายการ",
        "operationId": "getOrder",
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "อัปเดตสถานะคำสั่งซื้อ",
        "operationId": "updateOrderStatus",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "$ref": "#/components/schemas/OrderStatus"
                  },
                  "trackingNumber": {
                    "type": "string",
                    "example": "TH1234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "อัปเดตสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/coupons": {
      "get": {
        "tags": [
          "Coupons"
        ],
        "summary": "รายการคูปองทั้งหมด",
        "operationId": "listCoupons",
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Coupon"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Coupons"
        ],
        "summary": "ตรวจสอบรหัสคูปองกับยอดสั่งซื้อ",
        "operationId": "validateCoupon",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "DISCOUNT500"
                  },
                  "subtotal": {
                    "type": "number",
                    "example": 3500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ใช้คูปองได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Coupon"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "รหัสไม่ถูกต้อง หรือยอดไม่ถึงขั้นต่ำ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/api/orders/{id}/code": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "ORD-260814-A7F3C"
        }
      ],
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "ขอรหัส Steam Guard ของบัญชีที่ซื้อ",
        "description": "ขอได้ 3 รอบต่อคำสั่งซื้อ แต่ละรอบเปิดหน้าต่าง 60 วินาที ภายในหน้าต่างเดิมขอซ้ำได้โดยไม่นับรอบใหม่ (ไม่ต้องส่ง reason)",
        "operationId": "requestSteamGuardCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "ใส่เมื่อเปิดรอบใหม่"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ได้รหัส",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "valid_for_sec": {
                          "type": "integer"
                        },
                        "expiresInSec": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "ขอครบ 3 รอบแล้ว"
          },
          "404": {
            "description": "ไม่พบบัญชีเกมของคำสั่งซื้อนี้"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Cookie session ของ Supabase (ชื่อจริงคือ `sb-<project-ref>-auth-token`) เบราว์เซอร์ส่งให้อัตโนมัติหลังล็อกอิน"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token (JWT) สำหรับ API client ที่ไม่มี cookie jar"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "ยังไม่ได้ล็อกอิน หรือ token หมดอายุ/ไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "unauthorized",
              "message": "กรุณาเข้าสู่ระบบก่อนเรียกใช้งาน API"
            }
          }
        }
      },
      "BadRequest": {
        "description": "ข้อมูลที่ส่งมาไม่ครบหรือไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "ไม่พบข้อมูลตาม ID ที่ระบุ",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "ข้อผิดพลาดฝั่งเซิร์ฟเวอร์",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "CustomerInfo": {
        "type": "object",
        "required": [
          "name",
          "email",
          "phone",
          "address",
          "district",
          "province",
          "postalCode"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "คุณสมชาย ใจดี"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "example": "081-987-6543"
          },
          "address": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "postalCode": {
            "type": "string",
            "example": "10110"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "รอดำเนินการ",
          "กำลังจัดเตรียม",
          "จัดส่งแล้ว",
          "สำเร็จ",
          "ยกเลิก"
        ]
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "wallet",
          "promptpay",
          "credit_card",
          "bank_transfer",
          "cod"
        ]
      },
      "Order": {
        "type": "object",
        "required": [
          "id",
          "createdAt",
          "customer",
          "items",
          "subtotal",
          "discount",
          "shippingFee",
          "totalAmount",
          "status",
          "paymentMethod",
          "isPaid"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "ORD-260814-A7F3C"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-14 09:30"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerInfo"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "subtotal": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "shippingFee": {
            "type": "number"
          },
          "totalAmount": {
            "type": "number"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "isPaid": {
            "type": "boolean"
          },
          "trackingNumber": {
            "type": "string"
          },
          "couponCode": {
            "type": "string"
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
      "Coupon": {
        "type": "object",
        "required": [
          "code",
          "discountPercent",
          "minSpend",
          "description"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "DISCOUNT500"
          },
          "discountPercent": {
            "type": "number",
            "example": 15
          },
          "minSpend": {
            "type": "number",
            "example": 1000
          },
          "description": {
            "type": "string"
          },
          "freeShipping": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "description": "สำเนาข้อมูลสินค้า ณ เวลาที่สั่งซื้อ แก้สินค้าภายหลังไม่กระทบคำสั่งซื้อเดิม",
        "required": [
          "productId",
          "name",
          "unitPrice",
          "quantity"
        ],
        "properties": {
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "format": "uri"
          },
          "unitPrice": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "selectedColor": {
            "type": "string"
          }
        }
      },
...
    }
  }
}